- Ifc: enable double-precision arithmetics for the entire pipeline. This does fix bugs, but unfortunately not all if them.
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@1143 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/IFCUtil.cpp
@@ -269,7 +269,7 @@ void ConvertColor(aiColor4D& out, const IfcColourRgb& in)
 void ConvertColor(aiColor4D& out, const IfcColourOrFactor& in,ConversionData& conv,const aiColor4D* base)
 {
 	if (const EXPRESS::REAL* const r = in.ToPtr<EXPRESS::REAL>()) {
-		out.r = out.g = out.b = *r;
+		out.r = out.g = out.b = static_cast<float>(*r);
 		if(base) {
 			out.r *= static_cast<float>( base->r );
 			out.g *= static_cast<float>( base->g );
